refactor: move cluster aspect to related folder for cleaner
