+++ src/main/java/org/phong/zenflow/plugin/subdomain/nodes/builtin/integration/ai/base/tools/ToolRouter.java
@@ -1,4 +1,4 @@
-package org.phong.zenflow.plugin.subdomain.nodes.builtin.integration.ai.cluster;
+package org.phong.zenflow.plugin.subdomain.nodes.builtin.integration.ai.base.tools;
 
 import java.util.List;
