+++ src/main/java/org/phong/zenflow/plugin/subdomain/nodes/builtin/integration/ai/base/tools/tools.schema.json
