+++ src/test/java/org/phong/zenflow/plugin/subdomain/nodes/builtin/integration/ai/cluster/AiClusterComplexScenarioTest.java
@@ -12,6 +12,7 @@ import org.phong.zenflow.plugin.subdomain.nodes.builtin.integration.ai.base.AiEx
 import org.phong.zenflow.plugin.subdomain.nodes.builtin.integration.ai.base.AiToolRegistry;
 import org.phong.zenflow.plugin.subdomain.nodes.builtin.integration.ai.base.dto.AiExecutionRequest;
 import org.phong.zenflow.plugin.subdomain.nodes.builtin.integration.ai.base.factory.AiExecutionRequestFactory;
+import org.phong.zenflow.plugin.subdomain.nodes.builtin.integration.ai.base.tools.NodeToolCallback;
 import org.phong.zenflow.workflow.subdomain.context.ExecutionContext;
 import org.phong.zenflow.workflow.subdomain.engine.orchestrator.NodeExecutionOrchestrator;
 import org.phong.zenflow.workflow.subdomain.logging.core.NodeLogPublisher;
